Move ZmqService into separate module
There is a lot of code in the zmq_driver module and this is a cohesive whole
that can be removed. This also follows the Python SDK pattern of separating
these.
